refactor: 중복되는 캘린더 로직 훅으로 분리
